test: readPopularThemesByPeriod 테스트 추가
