+++ src/test/java/roomescape/controller/auth/AuthenticationControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.integration;
+package roomescape.controller.auth;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.hamcrest.Matchers.equalTo;
